refactor(DOMPurify): set related functions in separate file
