+++ front/src/middlewares/auth.js
@@ -1,24 +1,10 @@
 import { LOG_IN, UPDATE_USER, GET_USER, DELETE_TOKEN, DISCONNECT, UPDATE_SESSION, CHECK_TOKEN, SUBSCRIBE, UPDATE_PROFILE, REQUEST_SUCCESS } from '../actions'
-import DOMPurify from 'dompurify'
-
+import { cleanObject } from '../functions/DOMPurify'
 const auth = (store) => (next) => (action) => {
   const { email, password } = store.getState().user
   const back = store.getState().back.url
   const token = localStorage.getItem('token')
-  const clean = (dirty) => {
-    const cleant = DOMPurify.sanitize(dirty)
-    console.log(cleant)
-    return cleant
-  }
-  const cleanObject = (object = ({ login: 'login', password: 'password <script></script>' })) => {
-    const cleantObject = Object.entries(object).map(([key, value]) => {
-      const cleantValue = clean(value)
-      console.log(cleantValue)
-      return [key, cleantValue]
-    }).reduce((previousValue, currentValue) => ({ [previousValue[0]]: previousValue[1], [currentValue[0]]: currentValue[1] }))
 
-    return cleantObject
-  }
   switch (action.type) {
     case LOG_IN: {
       const login = {
